requirements.txt
# Django and core dependencies
Django==5.0.7
asgiref==3.8.1
sqlparse==0.5.1
typing-extensions==4.12.2

# Django REST Framework and extensions
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.2
django-ratelimit==4.2.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-decouple==3.8

# AWS/MinIO
boto3==1.34.140
botocore==1.34.140

# Image processing
Pillow==10.4.0

# Testing and development
Faker==25.8.0

# Cache
redis==5.0.7

# CORS (for development)
django-cors-headers==4.4.0

# Debug toolbar (for development)
django-debug-toolbar==4.4.6

# Security improvements
django-environ==0.11.2

# Additional utilities
django-extensions==3.2.3
celery==5.4.0
flower==2.0.1

# Validation
pydantic==2.8.2

# HTTP client
httpx==0.27.0

# Background tasks
celery[redis]==5.4.0
